+++ src/main/java/org/cric/back_office/user/repository/RefreshTokenRepository.java
@@ -11,7 +11,7 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
     
     Optional<RefreshToken> findByToken(String token);
     
-    Optional<RefreshToken> findByUserId(Integer userId);
+    Optional<RefreshToken> findByUserId(Long userId);
     
     void deleteByUserId(Integer userId);
